+++ src/DevTools.tsx
@@ -1,6 +1,6 @@
-import React from 'react';
+import React, { useState, useEffect, useRef } from 'react';
 
-import { Borders, DetectRowWrap } from './index';
+import { Borders, DetectRowWrapController } from './index';
 
 // See https://github.com/mobxjs/mobx-react-devtools/blob/6.1.1/src/Panel/styles/index.js
 const panel: React.CSSProperties = {
@@ -17,76 +17,63 @@ const panel: React.CSSProperties = {
 };
 
 export interface Props {
-  position?: {
-    top?: number;
-    right?: number;
-    bottom?: number;
-    left?: number;
+  rootRef: React.RefObject<HTMLElement>;
+  enableDetectRowWrap: boolean;
+  position: {
+    top: number;
+    right: number;
+    bottom: number;
+    left: number;
   };
 }
 
-interface State {
-  bordersDisplayed: boolean;
-  detectRowWrapEnabled: boolean;
+export function DevTools(props: Props) {
+  const { rootRef, enableDetectRowWrap, position } = props;
+
+  const [bordersDisplayed, setBordersDisplayed] = useState(Borders.displayed());
+  const [detectRowWrapEnabled, setDetectRowWrapEnabled] = useState(enableDetectRowWrap);
+
+  // See [Is there something like instance variables?](https://reactjs.org/docs/hooks-faq.html#is-there-something-like-instance-variables)
+  const detectRowWrap = useRef<DetectRowWrapController>();
+  useEffect(() => {
+    detectRowWrap.current = new DetectRowWrapController(rootRef.current!, enableDetectRowWrap);
+  }, [rootRef, enableDetectRowWrap]);
+
+  function handleToggleBordersChange() {
+    const displayed = Borders.toggle();
+    setBordersDisplayed(displayed);
+    const enabled = detectRowWrap.current!.run();
+    setDetectRowWrapEnabled(enabled);
+  }
+
+  function handleToggleDetectRowWrapChange() {
+    const enabled = detectRowWrap.current!.toggle();
+    setDetectRowWrapEnabled(enabled);
+  }
+
+  return (
+    <div style={{ ...panel, ...position }}>
+      <label title="Dotted line: growing block, solid line: fixed block">
+        <input type="checkbox" checked={bordersDisplayed} onChange={handleToggleBordersChange} />{' '}
+        Borders
+      </label>
+      {'  '}
+      <label title="Make previous and next blocks grow">
+        <input
+          type="checkbox"
+          checked={detectRowWrapEnabled}
+          onChange={handleToggleDetectRowWrapChange}
+        />{' '}
+        detectRowWrap()
+      </label>
+    </div>
+  );
 }
 
-export class DevTools extends React.Component<Props, State> {
-  static defaultProps: Props = {
-    position: {
-      // See https://github.com/mobxjs/mobx-react-devtools/blob/6.1.1/src/Panel/index.jsx#L40
-      top: -2,
-      right: 20
-    }
-  };
-
-  constructor(props: Props) {
-    super(props);
-
-    this.state = {
-      bordersDisplayed: Borders.displayed(),
-      detectRowWrapEnabled: DetectRowWrap.enabled
-    };
+DevTools.defaultProps = {
+  position: {
+    // See https://github.com/mobxjs/mobx-react-devtools/blob/6.1.1/src/Panel/index.jsx#L40
+    top: -2,
+    right: 20
   }
-
-  handleToggleBordersChange() {
-    const bordersDisplayed = Borders.toggle();
-    const detectRowWrapEnabled = DetectRowWrap.run();
-    this.setState({
-      bordersDisplayed,
-      detectRowWrapEnabled
-    });
-  }
-
-  handleToggleDetectRowWrapChange() {
-    const detectRowWrapEnabled = DetectRowWrap.toggle();
-    this.setState({
-      detectRowWrapEnabled
-    });
-  }
-
-  render() {
-    const { position } = this.props;
-
-    return (
-      <div style={{ ...panel, ...position }}>
-        <label title="Dotted line: growing block, solid line: fixed block">
-          <input
-            type="checkbox"
-            checked={this.state.bordersDisplayed}
-            onChange={this.handleToggleBordersChange.bind(this)}
-          />{' '}
-          Borders
-        </label>
-        {'  '}
-        <label title="Make previous and next blocks grow">
-          <input
-            type="checkbox"
-            checked={this.state.detectRowWrapEnabled}
-            onChange={this.handleToggleDetectRowWrapChange.bind(this)}
-          />{' '}
-          detectRowWrap()
-        </label>
-      </div>
-    );
-  }
-}
+};