+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Modernize build system
 - `$layout-item-margin-x` and `$layout-item-margin-y` instead of just `$layout-item-margin`
 - Bootstrap 4 example and demo
+- React hook `useDetectRowWrap()`
 
 ## v0.0.2 (2017/09/12)